pyproject.toml
[project]
name = "aiwf-eval-tinkering"
version = "0.1.0"
description = "Eval runner for AI Engineer World's Fair conversations"
requires-python = ">=3.12"
dependencies = [
  "pipecat-ai[google,openai,aws,azure]==0.0.97",
  "google-genai==1.51.0",
  "python-dotenv",
  "loguru",
  "sounddevice>=0.4.6",
  "soundfile>=0.12.1",
  "numpy>=1.26",
  "claude-agent-sdk>=0.1.8",
]

[project.scripts]
# Run with: `uv run record-turn-audio -- --start-index 0`
record-turn-audio = "scripts.record_turn_audio:main"
resample-turn-audio = "scripts.resample_turn_audio:main"

[build-system]
requires = ["hatchling>=1.20"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Ensure the `scripts` package (our CLI) is included in the wheel
packages = ["scripts"]

[tool.hatch.build.targets.sdist]
only-include = ["scripts", "scripts/record_turn_audio.py", "turns.py"]

# [tool.uv.sources]
# Use local checkout of pipecat instead of PyPI (commented out to use PyPI versions)
# "pipecat-ai" = { path = "vendor/pipecat" }
